validate signature when adding a local peer entry via chain maint Fixes #39 prevent runaway
