+++ src/diyims/scheduler.py
@@ -8,8 +8,9 @@ from diyims.logger_utils import get_logger, logger_server_main
 from diyims.config_utils import get_scheduler_config_dict
 from diyims.queue_server import queue_main
 from diyims.database_utils import reset_peer_table_status
-from diyims.general_utils import select_local_peer_and_update_metrics
+from diyims.peer_utils import select_local_peer_and_update_metrics
 from diyims.header_chain_utils import monitor_peer_publishing
+from diyims.peer_utils import monitor_peer_table_maint
 
 from multiprocessing import Process, set_start_method, freeze_support
 
@@ -47,6 +48,10 @@ def scheduler_main():
         sleep(int(scheduler_config_dict["submit_delay"]))
         monitor_peer_publishing_main_process.start()
         logger.debug("monitor_peer_publishing_main_process started.")
+        monitor_peer_table_maint_process = Process(target=monitor_peer_table_maint)
+        sleep(int(scheduler_config_dict["submit_delay"]))
+        monitor_peer_table_maint_process.start()
+        logger.debug("update metrics started.")
 
     if scheduler_config_dict["reset_enable"] == "True":
         reset_peer_table_status_process = Process(target=reset_peer_table_status)
@@ -134,6 +139,8 @@ def scheduler_main():
     if scheduler_config_dict["publish_enable"] == "Only":
         publish_main_process.join()
         monitor_peer_publishing_main_process.join()
+        monitor_peer_table_maint_process.join()
+
     if scheduler_config_dict["beacon_enable"] == "True":
         beacon_main_process.join()
         satisfy_main_process.join()
@@ -151,6 +158,7 @@ def scheduler_main():
         logger.info("issuing terminate of publish.")
         publish_main_process.terminate()
         monitor_peer_publishing_main_process.terminate()
+        monitor_peer_table_maint_process.terminate()
 
     if scheduler_config_dict["provider_enable"] == "True":
         logger.info("issuing terminate of logger_server_provider .")
